test(core): consolidate duplicate sandbox-policy and tunnel-service suites (#28806)
Consolidate parallel colocated and __tests__ suites for sandbox-policy
and tunnel-service. Merge unique assertions: direct/store gating with
both message variants (Tool requires + direct download), ServiceType.TUNNEL
call verification, and tunnelSlotIsFree true/false. Delete redundant
__tests__ files after verifying union preservation (4+7 vs 3+3/3+5).

Closes #28806
